MPP: withTimeout, delay in int number of milliseconds

+++ common/kotlinx-coroutines-core-common/src/main/kotlin/kotlinx/coroutines/experimental/CommonDelay.kt
@@ -2,4 +2,4 @@ package kotlinx.coroutines.experimental
 
 public expect interface Delay
 
-public expect suspend fun delay(time: Long)
+public expect suspend fun delay(time: Int)
